[Portal] Restructure the DCR tab into settings cards with a separate client list page
The Dynamic clients tab (renamed "Dynamic clients (DCR)") now hosts the
whole DCR configuration as Radix section cards: enable toggle with the
registration endpoint, a registered-clients count card with View all,
registration security (IAT management), and default client config with
a save bar. The paginated client list moves to its own page at
apps/dcr with a breadcrumb back to the tab. The standalone settings
screen is removed and its i18n keys renamed to DynamicClientsTab.*.

ref DEV-3801
