+++ portal/src/components/dynamic-clients/DynamicClientsTab.module.css
@@ -2,10 +2,10 @@
   @apply flex flex-col gap-4;
 }
 
-.header {
-  @apply flex flex-row items-start justify-between gap-4;
+.rootWithSaveBar {
+  padding-bottom: 5.5rem;
 }
 
-.headerText {
-  @apply flex min-w-0 flex-col gap-1;
+.saveBarClearance {
+  margin-bottom: 5.5rem;
 }
