+++ portal/src/locale-data/en.json
@@ -788,13 +788,16 @@
   "ApplicationsConfigurationScreen.title": "Applications",
   "ApplicationsConfigurationScreen.description": "Configure applications to use this project for authentication.",
   "ApplicationsConfigurationScreen.tab.applications": "Applications",
-  "ApplicationsConfigurationScreen.tab.dynamic-clients": "Dynamic clients",
+  "ApplicationsConfigurationScreen.tab.dynamic-clients": "Dynamic clients (DCR)",
+  "DynamicClientsTab.enable.toggle.label": "Enable dynamic client registration",
+  "DynamicClientsTab.enable.description": "Let OAuth clients, such as MCP clients, register themselves with this project at runtime via the registration endpoint (RFC 7591).",
+  "DynamicClientsTab.section.registration.title": "Dynamic client registration",
+  "DynamicClientsTab.clients.title": "Registered clients",
+  "DynamicClientsTab.clients.count": "{count} dynamic clients registered",
+  "DynamicClientsTab.clients.view-all": "View all",
+  "DynamicClientListScreen.title": "Dynamic clients",
   "DynamicClientsTab.registration-endpoint.label": "Registration endpoint",
   "DynamicClientsTab.quota": "{count} of {quota} dynamic clients used",
-  "DynamicClientsTab.settings-button": "Registration settings",
-  "DynamicClientsEmptyView.disabled.title": "Dynamic client registration is off",
-  "DynamicClientsEmptyView.disabled.description": "Let OAuth clients, such as MCP clients, register themselves with this project at runtime.",
-  "DynamicClientsEmptyView.disabled.cta": "Enable registration",
   "DynamicClientsEmptyView.empty.title": "No dynamic clients yet",
   "DynamicClientsEmptyView.empty.description": "Clients that register themselves through the registration endpoint appear here automatically.",
   "DynamicClientList.columns.name": "Name",
@@ -2476,25 +2479,20 @@
   "ResourceForm.name.description": "A human friendly name for display in the portal.",
   "ResourceForm.resourceURI.description": "An identifier of the resource. It need not be a publicly available URI. Authgear will not access the URI at all.<br></br>The value appears as the <code>aud</code> (audience) claim in the access token.",
   "ResourceForm.resourceURI.tooltip": "For example, use <code>https://api.myapp.com/v1</code> to represent your publicly available API server, or use <code>https://myapp.com/mcp/my-mcp</code> to represent a MCP server serving with stdio.",
-  "DynamicClientRegistrationScreen.title": "Dynamic client registration",
-  "DynamicClientRegistrationScreen.description": "Let OAuth clients register themselves with this project at runtime via the OAuth 2.0 Dynamic Client Registration protocol (RFC 7591). Registered clients appear in the Dynamic clients tab of Client Applications.",
-  "DynamicClientRegistrationScreen.enable.title": "Enable registration",
-  "DynamicClientRegistrationScreen.enable.toggle.label": "Enable dynamic client registration",
-  "DynamicClientRegistrationScreen.enable.toggle.description": "Expose the client registration endpoint so clients can register themselves with this project.",
-  "DynamicClientRegistrationScreen.security.title": "Registration security",
-  "DynamicClientRegistrationScreen.iat-required.toggle.label": "Require initial access token",
-  "DynamicClientRegistrationScreen.iat-required.toggle.description": "Only callers presenting a valid initial access token can register a client. Turning this off allows open registration: anyone can register a client with this project.",
-  "DynamicClientRegistrationScreen.open-registration.confirm.title": "Allow open registration?",
-  "DynamicClientRegistrationScreen.open-registration.confirm.description": "Without the initial access token requirement, anyone can register OAuth clients with this project. If your plan caps the number of dynamic clients, strangers can use up the cap, and registration stays blocked until an admin deletes clients. You can turn the requirement back on at any time.",
-  "DynamicClientRegistrationScreen.open-registration.confirm.confirm": "Allow open registration",
-  "DynamicClientRegistrationScreen.open-registration.confirm.cancel": "Cancel",
-  "DynamicClientRegistrationScreen.default-client-config.title": "Default client configuration",
-  "DynamicClientRegistrationScreen.default-client-config.description": "Token settings applied to every dynamically registered client. Leave a field empty to use the default shown.",
-  "DynamicClientRegistrationScreen.access-token-lifetime.label": "Access token lifetime (seconds)",
-  "DynamicClientRegistrationScreen.refresh-token-lifetime.label": "Refresh token lifetime (seconds)",
-  "DynamicClientRegistrationScreen.refresh-token-idle-timeout-enabled.label": "Invalidate refresh token after idling",
-  "DynamicClientRegistrationScreen.refresh-token-idle-timeout-enabled.description": "End a client's session when its refresh token has not been used for the idle timeout period.",
-  "DynamicClientRegistrationScreen.refresh-token-idle-timeout.label": "Refresh token idle timeout (seconds)",
+  "DynamicClientsTab.security.title": "Registration security",
+  "DynamicClientsTab.iat-required.toggle.label": "Require initial access token",
+  "DynamicClientsTab.iat-required.toggle.description": "Only callers presenting a valid initial access token can register a client. Turning this off allows open registration: anyone can register a client with this project.",
+  "DynamicClientsTab.open-registration.confirm.title": "Allow open registration?",
+  "DynamicClientsTab.open-registration.confirm.description": "Without the initial access token requirement, anyone can register OAuth clients with this project. If your plan caps the number of dynamic clients, strangers can use up the cap, and registration stays blocked until an admin deletes clients. You can turn the requirement back on at any time.",
+  "DynamicClientsTab.open-registration.confirm.confirm": "Allow open registration",
+  "DynamicClientsTab.open-registration.confirm.cancel": "Cancel",
+  "DynamicClientsTab.default-client-config.title": "Default client configuration",
+  "DynamicClientsTab.default-client-config.description": "Token settings applied to every dynamically registered client. Leave a field empty to use the default shown.",
+  "DynamicClientsTab.access-token-lifetime.label": "Access token lifetime (seconds)",
+  "DynamicClientsTab.refresh-token-lifetime.label": "Refresh token lifetime (seconds)",
+  "DynamicClientsTab.refresh-token-idle-timeout-enabled.label": "Invalidate refresh token after idling",
+  "DynamicClientsTab.refresh-token-idle-timeout-enabled.description": "End a client's session when its refresh token has not been used for the idle timeout period.",
+  "DynamicClientsTab.refresh-token-idle-timeout.label": "Refresh token idle timeout (seconds)",
   "InitialAccessTokenSection.title": "Initial access tokens",
   "InitialAccessTokenSection.description": "Pass an initial access token as a Bearer token when calling the registration endpoint. A first-party token registers clients that skip the consent screen.",
   "InitialAccessTokenSection.create": "Create token",
