--- portal/src/screens/dynamic-clients/DynamicClientRegistrationScreen.module.css
@@ -1,29 +1,0 @@
-.widget {
-  grid-column: 1 / span 9;
-
-  @apply tablet:col-span-full;
-}
-
-.pageHeader {
-  @apply flex flex-col gap-2;
-}
-
-.pageTitle {
-  margin: 0;
-  font-size: 20px;
-  line-height: 28px;
-  font-weight: 600;
-  color: var(--gray-12);
-}
-
-.pageDescription {
-  color: var(--gray-a11);
-}
-
-.contentWithSaveBar {
-  padding-bottom: 5.5rem;
-}
-
-.settingsCardSaveBarClearance {
-  margin-bottom: 5.5rem;
-}
